+++ pkg/adapter/adapter_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/rollkit/rollkit/types"
 
 	"github.com/rollkit/go-execution-abci/pkg/cometcompat"
+	execstore "github.com/rollkit/go-execution-abci/pkg/store"
 )
 
 func TestExecuteFiresEvents(t *testing.T) {
@@ -103,7 +104,7 @@ func TestExecuteFiresEvents(t *testing.T) {
 				AppHash:         []byte("apphash1"),
 			}
 
-			headerBz, err := cometcompat.SignaturePayloadProvider()(&header)
+			headerBz, err := cometcompat.SignaturePayloadProvider(adapter.Store)(&header)
 			require.NoError(t, err)
 
 			sig, err := privKey.Sign(headerBz)
@@ -118,7 +119,7 @@ func TestExecuteFiresEvents(t *testing.T) {
 				Signature: sig,
 			}
 			require.NoError(t, adapter.RollkitStore.SaveBlockData(ctx, signedHeader, &types.Data{Txs: make(types.Txs, 0)}, &sigT))
-			require.NoError(t, adapter.Store.SaveState(ctx, stateFixture()))
+			require.NoError(t, adapter.Store.SaveState(ctx, execstore.TestingStateFixture()))
 
 			// when
 			ctx = context.WithValue(ctx, types.SignedHeaderContextKey, signedHeader)
